Re-require strong-cluster-control so agent sees it
strong-supervisor plays a bit fast and loose with strong-agent, in that
it doesn't require it before strong-cluster-control (because until its
looked at the strong-cluster-control options, it can't evaluate its
command line, and doesn't know if it has to require strong-agent). This
adds an apparently side-effect-less require of strong-cluster-control
that allows the agent probes to see it is in use, and to enable the
cluster control panel.