Add --syslog option for direct syslog logging
When node-syslog can be loaded, the --syslog CLI flag enables syslog
logging for the supervisor and the workers. If node-syslog cannot be
loaded, the option is not listed in the help message and is ignored if
used.

In this mode, log line timestamping is suppressed because it is handled
by syslog(3).

Log messages from worker stderr are logged as LOG_CRIT while log lines
originating from stdout are logged as LOG_NOTICE.

Note for testing/debugging that Darwin defaults to only log LOG_NOTICE
and higher to disk.