logger: only use log transformer in clustered mode
There is no sane way to redirect the stdout/stderr of the current
process in node and that is where apps will log to when not clustered.
Since we cannot process the stdout/stderr of the app to add timestamps
and other tagging, we should disable our own log line prefixing for the
sake of consistency.

SLN-1060 - log cleanup