run: allow profiling to be explicitly started
It used to be profiling was only started when StrongOps was started.

This caused problems for metrics and other users of the agent, because
when agent is no started first, it doesn't behave well. In particular,
it doesn't monkey patch anything, causing metrics to be lacking probe
metrics, and agent traces and express records not be available.