test: fix test-runctl-clusterctl
We have to disable ipcctl in master because its parent is a bunch of
synchronous tests, which means the parent never reads from the IPC
channel node has set up for us. And since it never drains the buffer
it eventually fills up and process.send() becomes blocking. So now
the supervisor cluster master is blocked which basically causes
everything to go to hell. Now someone owes me a cider because I spent
a whole afternoon trying to figure out why this test started hanging
on the last command if I added more than 3 bytes to the status message.