trace-object: emit record as string, not object
Stringify the record object before forwarding it.  Avoids a number of
deserialize/reserialize steps down the chain.

There is still some overhead because the implicit JSON.stringify() and
JSON.parse() calls around process.send() escape and unescape the string
but that is cheaper than repeatedly (de)serializing the object graph.