make status message as complete as started message
Allows a controller to poll for the current status of a running master
and get a relatively complete picture even though it wasn't there to
receieve the original 'started' message. A concrete example of this is
when supervisor is running in a docker container and the PM instance
that spawned it was restarted.

Needed by strongloop/strong-pm#242