run: support simple start scripts as fallback
If the path given is not a file and is not a module path that can be
resolved to a file by `require()`, then look for the package.json
nearest to it and look for a scripts.start config. If it is there, and
if it is simple enough, use it to determine the path of the script to
run.

If the script is missing or not simple enough, fall back to the
original behaviour of checking for server.js, app.js, and index.js,
followed by node module resolution.