ws: reconnect WS ctl channel on error
If the WS control channel emits an error, it disconnects. It is up to
the client to re-connect to the server when this happens.

Depends on changes in strongloop/strong-control-channel#28 to improve
SCC's own handling of errors and disconnects, as well as improvements
to the testability of WS channels because of new events.