fix regression in handling of symlinked pwd
Fix a regression that was introduced by #144 that changed how symlinks
were being handled. Unfortunately, this affects pretty much every use
since strong-pm is the primary dependant of strong-supervisor and it
uses symlinks for all deployments. The reason the problem wasn't more
obvious is because the code path that triggers this only really happens
when the app's main script is in a subdirectory of the package.

The root cause of this is that some of the methods of resolving the
start script make use of the builtin require.resolve(). Because
require() has to know the physical location of the module it loads so
that it knows how to resolve files relative to it, it has to resolve
symlinks and return the real path.

To account for this we can check if the pwd given to the start command
resolver is a symlink and if it is reverse the symlink resolution
before passing back the result.

This should resolve strongloop/strong-pm#283