Remove support for running unclustered
The ability to run in unclustered mode is a relic of `slc run` being
used as a replacement for `node`, except with the strong-agent running
and reporting to the StrongOps service. The StrongOps service no longer
exists, this mode is no longer used, and complicates the code, so it is
now removed.