Remove startup variations related to StrongOps
StrongOps would start by default, and warn if it was unlicensed, but
could also be explicitly enabled, and explicitly disabled. This created
a complex triple mode startup process. Remove this complexity, its no
longer required, and the startup messages are confusing when they show
up in in Collective logs.

Also remove --detach, another source of startup complexity, which
created variations on the above modes and interacted with them in
difficult to understand ways.