test: trace tests work again with appmetrics 1.1.0
appmetrics 1.1.0 has no `.lrtime()` method when its not supported on a
platform, so strong-trace can dynamically determine how to get the
lo-res time. This allows us to run/test the trace tests on non-linux
platforms, which don't have lrtime. appmetrics < 1.1.0 always had a
`.lrtime()` method... but calling it would raise an uncaught exception.